updated storage.py and added logic of load and save to main file
