+++ CLI_Bot/AddressBook/addressbook_main.py
@@ -1,12 +1,12 @@
-from .addressbook_classes.addressbook_class import AddressBook
+# from .addressbook_classes.addressbook_class import AddressBook # Не потрібен при використанні завантаження з файлу
 from .handler import add_contact
 
 
-def addressbook_main():
+def addressbook_main(book):
 
     print("Hello! How i can help you?")
 
-    book = AddressBook()
+    # book = AddressBook() # Не потрібен при використанні завантаження з файлу
 
     while True:
